@@ -8,8 +8,8 @@ package com.portfolio.mgb.Security.Service;
 import com.portfolio.mgb.Security.Entity.Rol;
 import com.portfolio.mgb.Security.Enums.RolNombre;
 import com.portfolio.mgb.Security.Repository.iRolRepository;
+import jakarta.transaction.Transactional;
 import java.util.Optional;
-import javax.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
