@@ -7,8 +7,8 @@ package com.portfolio.mgb.Security.Service;
 
 import com.portfolio.mgb.Security.Entity.Usuario;
 import com.portfolio.mgb.Security.Repository.iUsuarioRepository;
+import jakarta.transaction.Transactional;
 import java.util.Optional;
-import javax.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
