@@ -7,9 +7,9 @@ package com.portfolio.mgb.Service;
 
 import com.portfolio.mgb.Entity.hys;
 import com.portfolio.mgb.Repository.Rhys;
+import jakarta.transaction.Transactional;
 import java.util.List;
 import java.util.Optional;
-import javax.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
